% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_functions.R
\name{ci_barplot_horizontal}
\alias{ci_barplot_horizontal}
\title{Create horizontal barplots.}
\usage{
ci_barplot_horizontal(
  df,
  my_category,
  my_title,
  my_facet = NULL,
  my_palette = "Paired"
)
}
\arguments{
\item{df}{Regular Dataframe}

\item{my_category}{Category that will be used for classification. Can be a
vector.}

\item{my_title}{String that will be used for Plot's title.}
}
\description{
Creates a horizontal barplot using ggplot.
}
