% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dishes_functions.R
\name{ci_dishratings}
\alias{ci_dishratings}
\title{Returns average ratings per dish.}
\usage{
ci_dishratings(dishes, dishratings)
}
\arguments{
\item{dishes}{Dataframe containing information regarding}

\item{dishratings}{Dataset containing dish ratings.}
}
\value{
Dishes dataframe with new columns, containing average ratings.
}
\description{
Reads dishrating dataset, calculates average ratings per dish
  and aggregates that information to dishes' dataset.
}
