% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{ci_get_data}
\alias{ci_get_data}
\title{Retrieves data from the tool and stores it a dataframe.}
\usage{
ci_get_data(json, filename = NULL, city = "slupsk")
}
\arguments{
\item{json}{String pointing to a JSON's URL from the tool.}

\item{filename}{String containing the filename to store a csv.}
}
\value{
Dataframe.
}
\description{
Retrieves json data from a url and converts it into a dataframe.
  Optionally, stores it in a CSV file.
}
