% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_functions.R
\name{ci_wf_prep_animals}
\alias{ci_wf_prep_animals}
\title{Reads and cleans animals' water footprint dataset.}
\usage{
ci_wf_prep_animals(
  filename = "Report48-Appendix-V.xlsx",
  sheetname = "App-V_WF_HS_SITC",
  skip_rows = 2,
  country = NULL
)
}
\arguments{
\item{filename}{String with the XLSX file name containing the water
footprint. File should be stored in data/data-raw folder.}

\item{sheetname}{String with the name of the Excel's sheet to be imported.}

\item{skip_rows}{Number containing the number of rows to be skipped when
importing the XLSX data.}

\item{country}{String with the country name we want to get water footprint
from.}
}
\value{
Dataframe in a tidy format, providing values for the selected country
  only.
}
\description{
Reads and cleans animals' water footprint dataset from
  https://waterfootprint.org/en/resources/waterstat/product-water-footprint-statistics/
   and converts it to tidy format.
}
