% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable_customisations.R
\name{rt_bar_cell}
\alias{rt_bar_cell}
\title{Renders a bar chart with a label on the left.}
\usage{
rt_bar_cell(value, variable, color, background = "#e1e1e1")
}
\arguments{
\item{value}{Cell's value.}

\item{variable}{dataframe's variable}

\item{color}{String containing color.}

\item{background}{String containing color to be used as background. Can be
set to NULL if no background is needed.}
}
\value{
Values aligned to the left and a barplot on its right.
}
\description{
Renders a bar chart with a label on the left.
}
