% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable_customisations.R
\name{rt_rating_icons}
\alias{rt_rating_icons}
\title{Displays ratings using icons.}
\usage{
rt_rating_icons(rating, icon = "star", color = "orange", max_rating = 5)
}
\arguments{
\item{rating}{Number, from cell's value.}

\item{icon}{String containing the name of the icon to be used. Icons defined
with shiny::icon()}

\item{color}{String containing color definition in hex format.}

\item{max_rating}{Number of maximum rating.}
}
\value{
An array of n icons and a tooltip showing the original value.
}
\description{
Converts numeric values into an array of icons, ideal for
  ratings.
}
