% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable_customisations.R
\name{rt_with_tooltip}
\alias{rt_with_tooltip}
\title{Changes header style when tooltips are enabled.}
\usage{
rt_with_tooltip(value, tooltip)
}
\arguments{
\item{value}{Cell's value.}

\item{dishes}{Dataframe containing information regarding}
}
\value{
A tooltip text, cursor change and header underline.
}
\description{
Adds text decoration to reactable's header, changes cursor type
  and shows a tooltip on header.
}
