% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_functions.R
\name{wf_data_prep}
\alias{wf_data_prep}
\title{Reads and cleans water footprint dataset.}
\usage{
wf_data_prep(filename, country = NULL)
}
\arguments{
\item{filename}{String with the XLSX file name containing the water
footprint. File should be stored in data/data-raw folder.}

\item{country}{String with the country name we want to get water footprint
from.}
}
\description{
Reads and cleans water footprint dataset from
  https://waterfootprint.org/en/resources/waterstat/product-water-footprint-statistics/
   and converts it to tidy format.
}
