% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_functions.R
\name{ci_dishratings_prep}
\alias{ci_dishratings_prep}
\title{Prepares dishrating dataset.}
\usage{
ci_dishratings_prep(filename)
}
\arguments{
\item{filename}{String pointing to file containing dish ratings' dataset.}
}
\value{
Clean and tidy dataset with dish ratings information.
}
\description{
Reads file containing dishrating dataset, cleans data and
  calculates average ratings per dish.
}
